adds artworkviewcontroller and more

+++ AFNetworkingLab/ArtWork.h
@@ -12,10 +12,10 @@
 @property (nonatomic, copy) NSString *title;
 @property (nonatomic, copy)NSString *artistName;
 @property (nonatomic, copy) NSString *mediumThumbUrl;
-@property (nonatomic) float price;
+@property (nonatomic) double price;
 
-+(id) artWorkWithTitle:(NSString*) title artistName: (NSString*) artistName mediumThumbUrl:(NSString*) mediumThumbUrl price: (float) price;
++(id) artWorkWithTitle:(NSString*) title artistName: (NSString*) artistName mediumThumbUrl:(NSString*) mediumThumbUrl price: (double) price;
 
--(id) initWithTitle:(NSString*) title artistName: (NSString*) artistName mediumThumbUrl:(NSString*) mediumThumbUrl price: (float) price;
+-(id) initWithTitle:(NSString*) title artistName: (NSString*) artistName mediumThumbUrl:(NSString*) mediumThumbUrl price: (double) price;
 
 @end
